modify calendar component
